Update new design for BalAlert (#3179)
* Update new design for BalAlert

* Add icon and color
