+++ src/pages/DexV2/Pool/Staking/ActionSteps.tsx
@@ -9,7 +9,7 @@ import { captureBalancerException, useErrorMsg } from 'lib/utils/errors'
 import useEthers from 'hooks/dex-v2/useEthers'
 import { postConfirmationDelay } from 'hooks/dex-v2/useTransactions'
 import { dateTimeLabelFor } from 'hooks/dex-v2/useTime'
-import { BalAlert } from 'pages/DexV2/common/BalAlert'
+import BalAlert from 'pages/DexV2/common/BalAlert'
 import { useSwapState } from 'state/dexV2/swap/useSwapState'
 import { setActionStates } from 'state/dexV2/swap'
 import BalBtn from 'pages/DexV2/common/popovers/BalBtn'
